+++ jest.config.js
@@ -9,6 +9,9 @@ module.exports = {
         "/docs/",
         "/docusaurus/"
     ],
+    roots: [
+        '<rootDir>/__tests__/',
+    ],
     testMatch: [
         "<rootDir>/__tests__/*.spec.js"
     ]